fix(examples): autonomous-pr — extract seeds to seed.ts, Agent terminology
Move the embedded scratch-repo seed constants (whose literal text contains
`entry:`/`defineAgent`) into a sibling seed.ts so the entry-schema check reads
the real `plan` entry step (inputSchema keys repoSlug/task/installCommand/
checkCommand) instead of the seeded hello-agent's `entry:"greet"`. Fix notes
copy: "workflow page" -> "Agents page". No graph/logic change.
